+++ EssentialFeed/Feed Presentation/en.lproj/Feed.strings
@@ -1,2 +1,3 @@
 "FEED_VIEW_TITLE" = "My Feed";
 "FEED_VIEW_CONNECTION_ERROR" = "Couldn't connect to server";
+
